% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.R
\name{num}
\alias{num}
\title{Convert a string expression into a number.}
\usage{
num(num)
}
\arguments{
\item{num}{character. The string expression to convert.}
}
\description{
Convert a string expression into a number.
}

